add alternative upscaler, more to come soon!
