+++ config/settings.ini
@@ -1,6 +1,6 @@
 [PathSettings]
 defaultOutputPath = $HOME/FSRImageVideoUpscaler/
-tmpPathLinux = /home/janis/
+tmpPathLinux = /tmp/
 tmpPathWindows = C:\temp
 
 [DevSettings]
